Fixed filtering issue and made a constants class
blinks were 0 because filtering was being applied so i reworked analysis class. There was a lot of redundant final constants so I put them in one class so it would be easier to manage. I applied the screen size once and then made an overloaded filterbyvalidity method to support filtering with a screen size