Fixed one off pattern issue in #55
Last pattern was not included in a sequence. Also set the minimums to always be positive to prevent breaking the looping constraints or index out of bounds. Changed the names of variables in Analysis class to make it more clear what sequences are being handled. Used List.of in Analysis class becuase it is structurally immutable (which is safer when not expecting the list to be modified in place)